chore(sessions): reconcile lifecycle quality gates
