+++ docs/internal/baselines/package-structure-baseline.json
@@ -2335,7 +2335,12 @@
     {
       "rule": "service-root-interface-count",
       "filePath": "pkg/services/factory_sessions",
-      "target": "pkg/services/factory_sessions/assembly_contract.go:InvocationService,pkg/services/factory_sessions/contracts.go:OpeningPresentationOwner,pkg/services/factory_sessions/execution_contract.go:DurableExecutionService,pkg/services/factory_sessions/execution_contract.go:SessionInventoryService,pkg/services/factory_sessions/service_contracts.go:LiveChangeAdmission,pkg/services/factory_sessions/service_contracts.go:LiveChangeApplication,pkg/services/factory_sessions/service_contracts.go:LiveChangeEventLog,pkg/services/factory_sessions/service_contracts.go:LiveChangePreflight,pkg/services/factory_sessions/service_contracts.go:LiveChangeService,pkg/services/factory_sessions/service_contracts.go:LiveControlService,pkg/services/factory_sessions/service_contracts.go:LiveResultService,pkg/services/factory_sessions/service_contracts.go:Service,pkg/services/factory_sessions/target_execution_contract.go:TargetExecutionService"
+      "target": "pkg/services/factory_sessions/assembly_contract.go:InvocationService,pkg/services/factory_sessions/contracts.go:OpeningPresentationOwner,pkg/services/factory_sessions/execution_contract.go:DurableExecutionService,pkg/services/factory_sessions/execution_contract.go:SessionInventoryService,pkg/services/factory_sessions/service_contracts.go:LiveChangeAdmission,pkg/services/factory_sessions/service_contracts.go:LiveChangeApplication,pkg/services/factory_sessions/service_contracts.go:LiveChangeEventLog,pkg/services/factory_sessions/service_contracts.go:LiveChangePreflight,pkg/services/factory_sessions/service_contracts.go:LiveChangeService,pkg/services/factory_sessions/service_contracts.go:LiveControlService,pkg/services/factory_sessions/service_contracts.go:LiveDeletionService,pkg/services/factory_sessions/service_contracts.go:LiveLifecycleControlService,pkg/services/factory_sessions/service_contracts.go:LiveResultService,pkg/services/factory_sessions/service_contracts.go:Service,pkg/services/factory_sessions/target_execution_contract.go:TargetExecutionService"
+    },
+    {
+      "rule": "service-root-interface-count",
+      "filePath": "pkg/services/factory_sessions/internal/services/live_runtime",
+      "target": "pkg/services/factory_sessions/internal/services/live_runtime/service.go:DeletionService,pkg/services/factory_sessions/internal/services/live_runtime/service.go:Service"
     },
     {
       "rule": "service-root-interface-count",
